Enforce `SCOREP_TOTAL_MEMORY` to be at most `UINT32_MAX` bytes.
Print warning if doing so. See ticket:1025.

Subversion: revision:9976

+++ src/measurement/include/SCOREP_Memory.h
@@ -78,8 +78,8 @@ typedef enum SCOREP_MemoryType
  * @param pageSize size of a single page in bytes, needs to be <= @a totalMemory
  */
 void
-SCOREP_Memory_Initialize( size_t totalMemory,
-                          size_t pageSize );
+SCOREP_Memory_Initialize( uint64_t totalMemory,
+                          uint64_t pageSize );
 
 
 /**
